change int to float on feed metrics

+++ NimbleRide/NimbleRide/FeedViewController.swift
@@ -59,9 +59,9 @@ class FeedViewController: UICollectionViewController, UICollectionViewDelegateFl
         nameDate.append(logo)
         cell.nameLabel.attributedText = nameDate
 
-        let avgSpeed = "Average Speed  - " + String(describing: Data[indexPath.row].avgSpeed as! Int) + " mph" + "\n"
+        let avgSpeed = "Average Speed  - " + String(describing: Data[indexPath.row].avgSpeed as! Float) + " mph" + "\n"
         let cals = "Calories Burned  - " + String(describing: Data[indexPath.row].calBurned as! Int) + "\n"
-        let dist = "Distance  - " + String(describing: Data[indexPath.row].distance as! Int) + " miles" + "\n"
+        let dist = "Distance  - " + String(describing: Data[indexPath.row].distance as! Float) + " miles" + "\n"
         let time = "Time  - " + Data[indexPath.row].time!
         cell.rideTextView.text = avgSpeed + cals + dist + time
